Added comments to code sample and corrected single-quotes
Added comments in lines 135 136 to point reader to helper.go to find implementation details for array(), concat(), and MakePair(), without which the sample code cannot work. (Took a while for me to realise that you weren't referencing an obscure built-in).

Changed single-quotes to double-quotes in lines 140 141 so we're correctly taking in strings instead of runes.